+++ common/src/main/java/dev/ftb/mods/ftbchunks/client/FTBChunksClientConfig.java
@@ -8,12 +8,11 @@ import dev.ftb.mods.ftbchunks.client.map.BiomeBlendMode;
 import dev.ftb.mods.ftbchunks.client.map.MapMode;
 import dev.ftb.mods.ftbchunks.client.minimap.MinimapComponentConfig;
 import dev.ftb.mods.ftbchunks.client.minimap.components.*;
-import dev.ftb.mods.ftbchunks.util.ChunkPosCustomYSetValue;
 import dev.ftb.mods.ftblibrary.config.manager.ConfigManager;
 import dev.ftb.mods.ftblibrary.snbt.config.*;
 import net.minecraft.resources.ResourceLocation;
 
-import java.util.*;
+import java.util.Collections;
 import java.util.stream.Stream;
 
 public interface FTBChunksClientConfig {